Phase6/src/auto/autonomous_driving_analysis/requirements.txt
# 크롤링
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
fake-useragent>=1.4.0

# 데이터 처리
pandas>=2.1.0
numpy>=1.24.0

# 텍스트 분석
konlpy>=0.6.0
wordcloud>=1.9.0

# 시각화 - 기본
matplotlib>=3.8.0
seaborn>=0.13.0

# 시각화 - 고급/인터랙티브
plotly>=5.18.0
kaleido>=0.2.1

# 시각화 - 추가
squarify>=0.4.3
networkx>=3.2.0
scipy>=1.11.0

# 유틸리티
tqdm>=4.66.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
